requirements.txt
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.2
exceptiongroup==1.1.1
iniconfig==2.0.0
lxml==4.9.2
lxml-stubs==0.4.0
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.2
PyYAML==6.0
soupsieve==2.4.1
tomli==2.0.1
types-beautifulsoup4==4.12.0.5
types-html5lib==1.1.11.14
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
